Various fixes to the gsg endpoint cli (#7)
- Apparently the seed can be larger than 32 bits. So use `i64` instead of `i32` for the response format.
- Improve the error message for when the response fails to deserialize into our expected format.
- Fix an incorrect `(duration, complexity)` match for PE, where it should be `(Duration::Long, Complexity::Complex)` instead of `(Duration::Long, Complexity::Simple)`.

+++ drg_mission_gen_gsg_endpoint_cli/src/deep_dive_response.rs
@@ -23,10 +23,10 @@ pub(crate) struct DeepDiveResponse {
     /// This may be a legacy seed used in older game versions. Does not seem to affect newer deep
     /// dives. Consider this unused.
     #[serde(rename = "Seed")]
-    _seed: i32,
+    _seed: i64,
     /// This is the seed used to generate both deep dive missions, which is the seed we're
     /// interested in.
-    pub(crate) seed_v2: i32,
+    pub(crate) seed_v2: i64,
     /// When does the dive expire?
     #[serde(rename = "ExpirationTime")]
     pub(crate) expiration_datetime: ExpirationDateTime,
